Fixes a race condition with D register, parameterises ROM in Top
B/c the ALU is combinatory, any time the ALU is configured for
an operation involving both inputs it actually computes it twice
b/c A and D updates twice per CPU cycle.

The original motivation for running A and D as 2x CPU clock is
b/c our RAM is synchrnous, its output is only valid on rising
clock edge and writes only happen on rising clock edge. In order
to implement D=M D needs to update twice b/c otherwise it won't see
the new value coming out of M b/c it loads on the same clock edge.

The solution is to double clock D (not need for A since A=M isn't
possible) only when the a-bit in a C-instruction is set. This is done
by using clk ^ (clk2x & ia) which produces an additional clock transition
for D when the a-bit (ia) is set allowing it to take on the value
from the ALU once and only once.

Lastly which program to load can now be specified as a parameter.