Removes RAM double clocking, addressM delay; adds pipelining in software.
*The code, as of this commit, runs correctly in simulation and on hw as
best as I can determine.*

The issue with sync vs async RAM is resolved here by introducing a nop
(0) before *any* RAM access (i.e. if M is present in a C instruction)
and for any writes to memory (i.e. if M is present in the LHS in a C
instruction) we insert a nop to give time for memory to commit. This
alone produces a working HACK platform.

In simulation there appears to be an issue with updating instruction on
negedge which causes a kind of delayed execution which doesn't seem to
affect the synthesised hardware. By moving instruction update to be
on the posedge the issue is resolved in simulation and in hw.