Fixes a huge mistake with how static segment was implemented
I didn't read the material carefully and assumed static was implemented like the temp segment
with a fixed base address of 16 which conflicts with how the assembler works, which auto-allocates
variables using the same fixed base address. The intended implementation is for each static
segment to be treated as access a special variable that is specific to the class/function, e.g.
foo.static.1, and let the assembler auto-assign a memory address.