Updates vm2asm.py to support multi-file translation and proper handling of file-level namespace...
which is required for statics to work as specified in the course.

Specifically the generated labels for static access is in the file namespace, e.g.
filename::static.STATIC0. This is necessary in order for different _function_ to access the same
static variables. Note that static in this case means _global_, not in the C case where a
function's static function is visible only to the function which is what I had implemented.

The namespace support is a bit janky. `label_in_namespace` really should be a function instead
of a property I think, but it works and I have other improvements to do.